+++ src/geeMines.c
@@ -69,7 +69,7 @@ POWNMENU _createMenu(){
             menubar_appendSubMenu(bar, sub, IDM_NEW, IDS_NEW, ITEM_STATE_DEFAULT, ITEM_STATUS_DEFAULT);
             menubar_appendItem(bar, IDM_START, IDS_START, ITEM_STATE_INACTIVE, ITEM_STATUS_DEFAULT);
 #ifdef _DEBUG_
-            menubar_appendItem(bar, IDM_DEBUG, IDS_DEBUG, ITEM_STATE_CHECKED, ITEM_STATUS_CHECKBOX);
+            menubar_appendItem(bar, IDM_DEBUG, IDS_DEBUG, ITEM_STATE_UNCHECKED, ITEM_STATUS_CHECKBOX);
 #endif // #ifdef _DEBUG_
             menubar_addItem(bar, MENU_POS_RIGHT, IDM_QUIT, IDS_QUIT, ITEM_STATE_DEFAULT, ITEM_STATUS_DEFAULT);
         }